+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+### Enhancements
+
+* Support `fac` in LEAN formula generation
+* Improve LEAN formula generation for `pow`
+
 ## v25.11.2
 
 ### Bugfixes